+++ core/Plugin/Manager.php
@@ -70,6 +70,7 @@ class Manager
 
     // These are always activated and cannot be deactivated
     protected $pluginToAlwaysActivate = array(
+        'BulkTracking',
         'CoreHome',
         'CoreUpdater',
         'CoreAdminHome',
@@ -1388,7 +1389,7 @@ class Manager
             return true;
         }
 
-        $hooks = $plugin->getListHooksRegistered();
+        $hooks = $plugin->registerEvents();
         $hookNames = array_keys($hooks);
         foreach ($hookNames as $name) {
             if (strpos($name, self::TRACKER_EVENT_PREFIX) === 0) {